# openapi.yaml
openapi: 3.0.3
info:
  title: Midas DVR / KLV API
  version: 1.0.0
  description: |
    HTTP API for source ingest, authoritative file workflows, HLS/KLV state,
    clips, snapshots, telemetry, and WebRTC signaling. File-backed operations
    use the private authoritative asset stored under mission-products/<productId>/source.
servers:
  - url: http://localhost:8090
    description: Default local server
...
paths:
...
  /ogc/processes/{processId}:
    get:
      tags: [OGC Processes]
      summary: Describe an OGC process
      parameters: [ { name: processId, in: path, required: true, schema: { type: string } } ]
      responses:
        '200': { description: Process description }
        '404': { $ref: '#/components/responses/Error' }
  /ogc/processes/{processId}/execution:
    post:
      tags: [OGC Processes]
      summary: Submit an asynchronous OGC process job
      parameters: [ { name: processId, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inputs]
              properties:
                inputs: { type: object, additionalProperties: true }
      responses:
        '201': { description: Job accepted; Location identifies its status resource }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /ogc/jobs/{jobId}:
    get:
      tags: [OGC Processes]
      summary: Read process job status
      parameters: [ { name: jobId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: Job status }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [OGC Processes]
      summary: Dismiss a job; an in-progress source provisioning job stops its source session
      parameters: [ { name: jobId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '204': { description: Job dismissed }
        '404': { $ref: '#/components/responses/Error' }
  /ogc/jobs/{jobId}/results:
    get:
      tags: [OGC Processes]
      summary: Read successful process outputs
      parameters: [ { name: jobId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: Process outputs }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /sources:
    get:
      tags: [Sources]
      summary: List active and transitional sources
      responses:
        '200':
          description: Source list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SourceRuntime' }
    post:
      tags: [Sources]
      summary: Start a stream URL or prepared authoritative file source
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartSourceRequest' }
      responses:
        '200':
          description: Source started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StartSourceResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
  /sources/{streamId}/state:
    get:
      tags: [Sources]
      summary: Get source lifecycle and processing state
      parameters: [ { $ref: '#/components/parameters/StreamId' } ]
      responses:
        '200':
          description: Source state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SourceRuntime' }
...
components:
  parameters:
    StreamId:
      name: streamId
      in: path
      required: true
      schema: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$' }
...
  responses:
    Error:
      description: Request could not be completed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties: { ok: { type: boolean, example: false }, error: { type: string } }
...
    StartSourceRequest:
      type: object
      required: [sourceType]
      properties:
        streamId: { type: string, readOnly: true, description: Service-generated backing ID returned from source allocation. }
        sourceType: { type: string, enum: [stream, file] }
        inputUrl: { type: string, description: Required for stream sources }
        assetId: { type: string, description: Required for file sources }
        hlsMode: { type: string, enum: [passthrough, abr], default: passthrough }
        webRtcMode: { type: string, enum: [auto, copy, transcode], default: auto }
        hlsSegmentSeconds: { type: number, default: 5 }
        vttSegmentSeconds: { type: number, default: 5 }
        maxCuesPerSecond: { type: number, default: 10 }
        minCueDurSec: { type: number, default: 0.1 }
        maxCueDurSec: { type: number, default: 0.5 }
    StartSourceResponse:
      type: object
      properties:
        ok: { type: boolean }
        streamId: { type: string }
        productId: { type: string, format: uuid, nullable: true, description: Published product ID; file sources expose it when ready. }
        hlsMasterUrl: { type: string }
        subtitlesUrl: { type: string }
        state: { $ref: '#/components/schemas/SourceRuntime' }
...
    SourceRuntime:
      type: object
      properties:
        streamId: { type: string }
        productId: { type: string, format: uuid, nullable: true }
        state: { type: string, enum: [stopped, starting, running, degraded, finalizing, ready, stopping, error] }
        sourceType: { type: string, enum: [stream, file] }
        running: { type: boolean }
        durationSeconds: { type: number, nullable: true }
        availableClipEndSeconds: { type: number, nullable: true, description: For a file source still packaging, the exclusive end of fully published browser HLS segments. It equals the full duration once ready. }
        availableClipSegmentCount: { type: integer, nullable: true, description: Number of fully published browser HLS segments used for the current clip boundary. }
        progressPercent: { type: number, nullable: true }
        processedSeconds: { type: number, nullable: true }
        lastError: { type: string, nullable: true }
...